# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Katalog Produk Global API",
        "description": "Dokumentasi API untuk Modul 1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/catalog/products/autocomplete": {
            "get": {
                "tags": [
                    "Module 2 - Search Optimization"
                ],
                "summary": "Autocomplete product names",
                "description": "Return product suggestions for autocomplete (id, name, slug) based on prefix match.",
                "operationId": "autocompleteProducts",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search prefix",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "sma"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/catalog/products/facets": {
            "get": {
                "tags": [
                    "Module 2 - Search Optimization"
                ],
                "summary": "Get facets for search UI (category counts, price ranges)",
                "description": "Return category counts and predefined price-range counts to drive UI filters.",
                "operationId": "productFacets",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Optional keyword to scope facets",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/catalog/products/bulk-search": {
            "post": {
                "tags": [
                    "Module 2 - Search Optimization"
                ],
                "summary": "Bulk lookup products by ids",
                "description": "Return multiple products by ID in a single request (useful for batch calls).",
                "operationId": "bulkSearchProducts",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/catalog/products/suggest": {
            "get": {
                "tags": [
                    "Module 2 - Search Optimization"
                ],
                "summary": "Suggest related products",
                "description": "Return related products (same category or similar) for recommendation widget.",
                "operationId": "suggestProducts",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Source product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/api/v1/internal/catalog/categories": {
            "post": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Create new category",
                "operationId": "storeCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Category"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "new-category"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/internal/catalog/categories/{id}": {
            "put": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Update category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Category"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Delete category",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}